refactor: reorganize components
