+++ src/app/layout.js
@@ -1,7 +1,7 @@
 import "./globals.css";
-import { AuthProvider } from "./AuthContext";
-import AuthGuard from "./AuthGuard";
-import HeaderBar from "./HeaderBar";
+import { AuthProvider } from "@/components/auth/AuthContext";
+import AuthGuard from "@/components/auth/AuthGuard";
+import HeaderBar from "@/components/HeaderBar";
 import { Mochiy_Pop_One, Quicksand } from "next/font/google";
 
 const mochiy = Mochiy_Pop_One({ weight: "400", subsets: ["latin"], variable: "--font-mochiy" });
