+++ src/app/page.js
@@ -1,4 +1,4 @@
-import DashboardContent from "./DashboardContent";
+import DashboardContent from "@/components/DashboardContent";
 export default function Home() {
   return <DashboardContent />;
 }
