+++ src/app/shopping/page.js
@@ -3,7 +3,7 @@ import { useState, useEffect } from "react";
 import Link from "next/link";
 import { db } from "@/lib/firebase";
 import { collection, getDocs, doc, getDoc } from "firebase/firestore";
-import { useAuth } from "../AuthContext";
+import { useAuth } from "@/components/auth/AuthContext";
 
 export default function ShoppingPage() {
   const { user } = useAuth();
